+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # @project-version@ (@build-time@)
 
 * Fixed a bug with Boss Mods widget that occurred because a deprecated value for vertical alignment of text was used [Comment #7295].
+* Fixed a bug with off-tank detection in Cata Classic that occurred because the wrong API for raid roles was used. 
 * Updated TOC version for Patch 10.2.7 for WoW Mainline.
 * Upgrade integrated libraries (Ace3 Release-r1341).